# src/main/resources/openapi/api.yml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Delegation Service
tags:
  - name: "Comment"
  - name: "Checkpoint"
  - name: "CheckpointV2"
  - name: "Delegation"
paths:
  /comments:
    post:
      summary: Creating new comment
      operationId: addComment
      tags:
        - "Comment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

    get:
      summary: Retrieving all comments from database
      tags:
        - "Comment"
      operationId: getAllComments
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
  /comments/comment/{id}/:
    get:
      summary: Retrieving comment from database by id
      tags:
        - "Comment"
      operationId: getCommentsById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Comment"
    delete:
      summary: Deleting comment by id
      tags:
        - "Comment"
      operationId: deleteCommentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Comment deleted'
    put:
      summary: Updating comment by id
      tags:
        - "Comment"
      operationId: updateCommentById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Comment updated'
  /comments/v2:
    post:
      summary: Creating new comment
      operationId: addCommentV2
      tags:
        - "Comment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentV2"
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentV2"

    get:
      summary: Retrieving all comments from database
      tags:
        - "Comment"
      operationId: getAllCommentsV2
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentV2"
  /comments/v2/comment/{id}/:
    get:
      summary: Retrieving comment from database by id
      tags:
        - "Comment"
      operationId: getCommentsByIdV2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/CommentV2"
    delete:
      summary: Deleting comment by id
      tags:
        - "Comment"
      operationId: deleteCommentByIdV2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Comment deleted'
    put:
      summary: Updating comment by id
      tags:
        - "Comment"
      operationId: updateCommentByIdV2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentV2"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Comment updated'







  /checkpoints/checkpoint/{delegationId}:
    get:
      summary: Get checkpoints
      tags:
        - "Checkpoint"
      operationId: getCheckpoints
      parameters:
        - name: delegationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkpoint"
  /checkpoints/checkpoint/{id}:
    put:
      summary: Updating an existing checkpoint
      tags:
        - "Checkpoint"
      operationId: update
      requestBody:
        description: Update an existing checkpoint by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Checkpoint'
        required: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Checkpoint set as checked'
    delete:
      summary: Deleting a checkpoint
      tags:
        - "Checkpoint"
      operationId: deleteCheckpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Checkpoint deleted'
        '400':
          description: 'Invalid id'
    patch:
      summary: Updating part of checkpoint
      tags:
        - "Checkpoint"
      operationId: changeStatus
      requestBody:
        description: Update checkpoint with properties to be changed
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/Checkpoint'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Updated
  /checkpoints/checkpoint/{checkpointId}/comment:
    get:
      summary: Get checkpoint comment
      tags:
        - "Checkpoint"
      operationId: getCheckpointComment
      parameters:
        - name: checkpointId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentToCheckpoint'
    post:
      summary: Post checkpoint comment
      tags:
        - "Checkpoint"
      operationId: add
      requestBody:
        description: Add a comment to the checkpoint
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentToCheckpoint'
        required: true
      parameters:
        - name: checkpointId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
  /checkpoints/checkpoint/{checkpointId}/comments:
    get:
      summary: Get checkpoint latest comment
      tags:
        - "Checkpoint"
      operationId: getCheckpointComments
      parameters:
        - name: checkpointId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/CommentToCheckpoint'
  /checkpoints/v2/checkpoint/{delegationId}:
    get:
      summary: Get checkpoints
      tags:
        - "CheckpointV2"
      operationId: getCheckpointsV2
      parameters:
        - name: delegationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckpointV2"
  /checkpoints/v2/checkpoint/{id}:
    put:
      summary: Updated an existing checkpoint
      tags:
        - "CheckpointV2"
      operationId: updateV2
      requestBody:
        description: Update an existing checkpoint by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckpointV2'
        required: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Checkpoint set as checked'
    delete:
      summary: Deleting a checkpoint
      tags:
        - "CheckpointV2"
      operationId: deleteCheckpointV2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Checkpoint deleted'
        '400':
          description: 'Invalid id'
    patch:
      summary: Updating part of checkpoint
      tags:
        - "CheckpointV2"
      operationId: changeStatusV2
      requestBody:
        description: Update checkpoint with properties to be changed
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CheckpointV2'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Updated











  /delegations:
    post:
      tags:
        - "Delegation"
      summary: Create a new delegation
      description: Create a new delegation
      operationId: createDelegation
      requestBody:
        description: Create a new delegation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Delegation'
        required: true
      responses:
        '200':
          description: Successful operation
        '405':
          description: Validation exception
    # get listy delegacji
  /delegations/{delegationId}:
    put:
      tags:
        - "Delegation"
      summary: Update an existing delegation by id
      description: Update an existing delegation by id
      operationId: updateDelegation
      requestBody:
        description: Update an existing delegation by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Delegation'
        required: true
      parameters:
        - name: delegationId
          in: path
          description: ID of delegation
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Delegation not found
        '405':
          description: Validation exception
    get:
      tags:
        - "Delegation"
      summary: Get a delegation by id
      description: Get a delegation by id
      operationId: getDelegation
      parameters:
        - name: delegationId
          in: path
          description: ID of delegation
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delegation'
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Delegation not found
    delete:
      tags:
        - "Delegation"
      summary: Delete a delegation by id
      description: Delete a delegation by id
      operationId: deleteDelegation
      parameters:
        - name: delegationId
          in: path
          description: Delegation to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied

components:
  schemas:
    Comment:
      description: This is a comment model
      type: object
      properties:
        id:
          description: The id of comment
          type: integer
          format: int64
          example: 16485
        delegationId:
          description: The id of delegation
          type: integer
          format: int64
          example: 15432
        parentCommentId:
          description: The id of the parent comment
          type: integer
          format: int64
          example: 16485
        author:
          description: It's an author of the comment
          type: string
          example: Martin Bobo
        text:
          type: string
          example: The text of the comment
        createdDate:
          type: string
          format: date-time
    CommentV2:
      description: This is a comment model
      type: object
      properties:
        id:
          description: The id of comment
          type: integer
          format: int64
          example: 16485
        delegationId:
          description: The id of delegation
          type: integer
          format: int64
          example: 15432
        parentCommentId:
          description: The id of the parent comment
          type: integer
          format: int64
          example: 16485
        author:
          description: It's an author of the comment
          type: string
          example: Martin Bobo
        commentTitle:
          type: string
          example: The text of the comment
        commentBody:
          type: string
          example: The text of the comment
        createdDate:
          type: string
          format: date-time




    Checkpoint:
      description: This is an element of checkpoints List
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 16425
        masterDataCheckpointId:
          type: integer
          format: int64
          example: 231
        isChecked:
          type: boolean
          example: false
        description:
          type: string
          example: Checkpoint name
        comment:
          type: string
          example: Checkpoint comment
    CheckpointV2:
      description: This is an element of checkpoints List
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 16425
        masterDataCheckpointId:
          type: integer
          format: int64
          example: 231
        isChecked:
          type: boolean
          example: false
        description:
          type: string
          example: Checkpoint name
        comment:
          type: array
          items:
            type: string
    CommentText:
      type: string
    CommentToCheckpoint:
      properties:
        id:
          type: integer
          format: int64
          example: 9
        comment:
          type: string
          example: Comment text
        checkpointId:
          type: integer
          format: int64
          example: 1






    Delegation:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 15432
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        fullname:
          type: string
          example: Martin Bobo
        location:
          type: string
          example: Berlin, Germany
        description:
          type: string
          example: Delegation description
    DelegationV2:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 15432
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        fullname:
          type: string
          example: Martin Bobo
        city:
          type: string
          example: Berlin
        country:
          type: string
          example: Germany
        description:
          type: string
          example: Delegation description